[Refactor] #95 - statisticsTabItem graph 제거 수술 실시
